+++ src/main/java/racingcar/model/Cars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.Collections;
 import java.util.HashSet;
