+++ src/main/java/racingcar/model/Engine.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public interface Engine {
         public abstract int active();
