+++ src/main/java/racingcar/model/RacingField.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
